Prolaze svi testovi!

+++ src/test/java/com/example/demo/DemoApplicationTests.java
@@ -2,38 +2,28 @@ package com.example.demo;
 
 import com.example.demo.model.AdminAddUser;
 import com.example.demo.model.AdminUserUpdate;
+import com.example.demo.model.Group;
 import com.example.demo.model.User;
 import com.example.demo.repo.*;
 import com.example.demo.service.FileService;
 import com.example.demo.service.GroupService;
 import com.example.demo.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
+import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.mock.web.MockHttpSession;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-
-import java.util.Optional;
-
-import static org.springframework.security.config.http.MatcherType.mvc;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
+@WithMockUser("admin")
+@Transactional
+@Rollback
 class DemoApplicationTests {
 	@Autowired
 	RoleRepository roleRepository;
@@ -57,32 +47,17 @@ class DemoApplicationTests {
 	UserGroupRepository userGroupRepository;
 	@Autowired
 	private HttpServletRequest request;
-/*	@Autowired
-	private MockMvc mockMvc;
 
-	void loginAdmin() throws Exception {
-		MockHttpSession session = new MockHttpSession();
-
-		Optional<User> user = userRepository.findById(1);
-
-		session.setAttribute("user", user.get());
-
-		MockHttpServletRequestBuilder builder = MockMvcRequestBuilders.get("/api/users")
-				.session(session);
-		this.mockMvc.perform(builder)
-				.andExpect(MockMvcResultMatchers.status()
-						.isOk());
-
-	}*/
 	@Test
 	void roles() {
 		Assertions.assertEquals(roleRepository.findAll().size(), 3);
 	}
 
 	@Test
+
 	void addGroup() {
-		groupService.addGroup("Managers", null);
-		Assertions.assertTrue(groupRepository.findByName("Managers") != null);
+		groupService.addGroup("Praktikanti", null);
+		Assertions.assertTrue(groupRepository.findByName("Praktikanti") != null);
 	}
 
 
@@ -101,8 +76,8 @@ class DemoApplicationTests {
 
 	@Test
 	public void addUserInGroup() {
-		groupService.addUserInGroup(1, 10);
-		Assertions.assertNotNull(userGroupRepository.findByGroupUser(1, 10));
+		groupService.addUserInGroup(7, 10);
+		Assertions.assertNotNull(userGroupRepository.findByGroupUser(7, 10));
 	}
 
 	@Test
@@ -122,13 +97,15 @@ class DemoApplicationTests {
 
 	@Test
 	public void deleteUser() {
-		userService.deleteUser(14);
+
+		if(userRepository.findById(14).isPresent()) userService.deleteUser(14);
 		Assertions.assertFalse(userRepository.findById(14).isPresent());
 	}
 
 	@Test
 	public void addUser() throws Exception {
-		//loginAdmin();
+		User user = userRepository.findByUsername("Selma13");
+		if(user != null) userRepository.deleteById(user.getId());
 		var userAdd = new AdminAddUser();
 		userAdd.setRole(2);
 		userAdd.setEmail("lealalalal5543@gmail.com");
@@ -146,8 +123,8 @@ class DemoApplicationTests {
 		userUpdate.setPassword("novipassword5");
 		userUpdate.setRole(2);
 		userUpdate.setUsername("Maja");
-		userService.updateUser(userUpdate, 14);
-		Assertions.assertEquals(userUpdate.getActive(), userRepository.findById(14).get().isActive());
+		userService.updateUser(userUpdate, 13);
+		Assertions.assertEquals(userUpdate.getActive(), userRepository.findById(13).get().isActive());
 	}
 /*
 	@Test
@@ -156,34 +133,35 @@ class DemoApplicationTests {
 	}*/
 
 	@Test
-	public void deleteFile() {
+	public void deleteNonExistingFile() {
 		Assertions.assertThrows(RuntimeException.class, () -> fileService.deleteFile(-1));
 	}
 
 	@Test
 	public void updateFile() {
-		var file = fileRepository.findById(34);
-		fileService.updateFile(34, file.get().getTitle(), file.get().getDescription(), file.get().getExpiry(), 20);
-		Assertions.assertEquals(20, fileRepository.findById(34).get().getMaxDownload());
+		var file = fileRepository.findById(38);
+		fileService.updateFile(38, file.get().getTitle(), file.get().getDescription(), file.get().getExpiry(), 20);
+		Assertions.assertEquals(20, fileRepository.findById(38).get().getMaxDownload());
 
 	}
 
 	@Test
+	@Rollback
 	public void addGroupInFile() {
-
-		fileService.addGroupInFile(34, 7);
-		Assertions.assertTrue(fileRepository.findById(34).isPresent());
-		Assertions.assertNotNull(fileGroupRepository.findByFileGroup(34, 7));
+		Group group = new Group();
+		group.setName("New");
+		fileService.addGroupInFile(38, 7);
+		Assertions.assertTrue(fileRepository.findById(38).isPresent());
+		Assertions.assertNotNull(fileGroupRepository.findByFileGroup(38, 7));
 
 	}
 
 	@Test
+	@Rollback
 	public void addUserInFile() {
-
-		fileService.addUserInFile(34, 10);
-		Assertions.assertTrue(fileRepository.findById(34).isPresent());
-		Assertions.assertNotNull(fileUserRepository.findByFileUser(34, 10));
-
+		fileService.addUserInFile(39, 13);
+		Assertions.assertTrue(fileRepository.findById(39).isPresent());
+		Assertions.assertNotNull(fileUserRepository.findByFileUser(39, 13));
 	}
 
 }
